+++ src/main/java/dao/FilesDAO.java
@@ -29,17 +29,6 @@ public class FilesDAO {
 	}
 	
 	
-	public void updateProfileImage(String userid, String sysName) throws Exception {
-//		회원이 선택한 프로필 이미지 정보를 DB의 회원 테이블에 저장함
-		
-		String sql = "UPDATE member SET profile_img = ? WHERE userid = ?";
-		try (Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql)) {
-			pstat.setString(1, "/image/mypage_image/" + sysName);
-            pstat.setString(2, userid);
-            pstat.executeUpdate();
-		}
-	}
